+++ crates/stella-script/src/tests/themes/particles.rs
@@ -86,10 +86,8 @@ fn theme_particle_spawner_updates_parallax_fields_and_draws_before_layer() {
     assert_eq!(sprites, ["THEME_PARTICLE", "THEME_LAYER"]);
     let particle = &bridge.commands[0];
     assert!(!particle.world_space);
-    let projected_x =
-        (particle.state.translate_x as f32 + particle.x as f32) * particle.state.scale_x as f32;
-    let projected_y =
-        (particle.state.translate_y as f32 + particle.y as f32) * particle.state.scale_y as f32;
+    let projected_x = (particle.state.translate_x + particle.x) * particle.state.scale_x;
+    let projected_y = (particle.state.translate_y + particle.y) * particle.state.scale_y;
     assert_eq!(projected_x, 128.381_26_f32);
     assert_eq!(projected_y, 102.006_25_f32);
     assert_eq!(particle.state.scale_x, 23.75);
